Unify log commands, completions
